refactor: move editor drawing logic to EditorPainter
